pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ainap"
version = "1.0.0"
description = "AI News Auto Publisher - AIニュース自動収集・分析・WordPress投稿システム"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "anthropic>=0.40",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "unidecode>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
